+++ packages/sanity/src/_singletons/structure/comments/intent/CommentsIntentContext.ts
@@ -1,7 +1,10 @@
 import {createContext} from 'react'
 
-import {type CommentsIntentContextValue} from './types'
+import {CommentsIntentContextValue} from '../../../../core/comments/context/intent/types'
 
+/**
+ * @internal
+ */
 export const CommentsIntentContext = createContext<CommentsIntentContextValue | undefined>(
   undefined,
 )
