+++ packages/sanity/src/_singletons/structure/comments/onboarding/CommentsOnboardingContext.ts
@@ -1,5 +1,7 @@
 import {createContext} from 'react'
+import {CommentsOnboardingContextValue} from '../../../../core/comments/context/onboarding/types'
 
-import {type CommentsOnboardingContextValue} from './types'
-
+/**
+ * @internal
+ */
 export const CommentsOnboardingContext = createContext<CommentsOnboardingContextValue | null>(null)
