+++ packages/sanity/src/core/comments/context/onboarding/index.ts
@@ -1,2 +1,1 @@
-export * from './CommentsOnboardingContext'
 export * from './CommentsOnboardingProvider'
