+++ packages/sanity/src/core/comments/context/upsell/CommentsUpsellProvider.tsx
@@ -1,6 +1,7 @@
 import {useTelemetry} from '@sanity/telemetry/react'
 import {template} from 'lodash'
 import {useCallback, useEffect, useMemo, useState} from 'react'
+import {CommentsUpsellContext} from 'sanity/_singletons'
 
 import {useClient, useProjectId} from '../../../hooks'
 import {
@@ -13,7 +14,6 @@ import {
 import {type UpsellData} from '../../../studio/upsell/types'
 import {UpsellDialog} from '../../../studio/upsell/UpsellDialog'
 import {DEFAULT_STUDIO_CLIENT_OPTIONS} from '../../../studioClient'
-import {CommentsUpsellContext} from './CommentsUpsellContext'
 import {type CommentsUpsellContextValue} from './types'
 
 const FEATURE = 'comments'
