+++ packages/sanity/src/core/comments/hooks/useCommentsSelectedPath.ts
@@ -1,6 +1,7 @@
 import {useContext} from 'react'
+import {CommentsSelectedPathContext} from 'sanity/_singletons'
 
-import {CommentsSelectedPathContext, type CommentsSelectedPathContextValue} from '../context'
+import {type CommentsSelectedPathContextValue} from '../context'
 
 export function useCommentsSelectedPath(): CommentsSelectedPathContextValue {
   const ctx = useContext(CommentsSelectedPathContext)
